net: Always add IP4 packet reassembly
